Validate pet IDs and preload edit form
Implemented validation for pet IDs during edit and delete operations to ensure data integrity. Extended validation to pet editing, pairing it with the existing form validation mechanism. Prepopulated the pet edit form with existing data for a better user experience. Refactored related action functions to support these validations and provide more informative error messages on failure.

+++ src/actions/actions.ts
@@ -2,13 +2,12 @@
 
 import prisma from "@/lib/db";
 import { PetEssentials } from "@/lib/types";
-import { petFormSchema } from "@/lib/validations";
+import { petFormSchema, petIdSchema } from "@/lib/validations";
 import { Pet } from "@prisma/client";
 import { revalidatePath } from "next/cache";
 
 export async function addPet(pet: unknown) {
   const validatedPet = petFormSchema.safeParse(pet);
-
   if (!validatedPet.success) {
     return {
       message: validatedPet.error.errors[0].message,
@@ -29,12 +28,21 @@ export async function addPet(pet: unknown) {
 }
 
 export async function editPet(petId: unknown, newPetData: unknown) {
+  const validatedPetId = petIdSchema.safeParse(petId);
+  const validatedPet = petFormSchema.safeParse(newPetData);
+
+  if (!validatedPetId.success || !validatedPet.success) {
+    return {
+      message: "Invalid pet data",
+    };
+  }
+
   try {
     await prisma.pet.update({
       where: {
-        id: petId,
+        id: validatedPetId.data,
       },
-      data: newPetData,
+      data: validatedPet.data,
     });
   } catch (error) {
     return {
@@ -46,10 +54,16 @@ export async function editPet(petId: unknown, newPetData: unknown) {
 }
 
 export async function deletePet(petId: unknown) {
+  const validatedPetId = petIdSchema.safeParse(petId);
+  if (!validatedPetId.success) {
+    return {
+      message: "Invalid pet data",
+    };
+  }
   try {
     await prisma.pet.delete({
       where: {
-        id: petId,
+        id: validatedPetId.data,
       },
     });
   } catch (error) {
